Fix filter insertion

+++ src/cuckoo_hashtable.cc
@@ -4,15 +4,15 @@
 
 namespace cuckoofilter {
 
-bool CuckooHashTable::Insert(uint16_t fingerprint, int bucket_idx) {
+bool CuckooHashTable::Insert(uint32_t fingerprint, int bucket_idx) {
   if (buckets_[bucket_idx].size() >= bucket_size) {
     return false;
   }
   return buckets_[bucket_idx].insert(fingerprint).second;
 }
 
-bool CuckooHashTable::Remove(uint16_t fingerprint, int bucket_idx) {
-  std::unordered_set<uint16_t> bucket = buckets_[bucket_idx];
+bool CuckooHashTable::Remove(uint32_t fingerprint, int bucket_idx) {
+  std::unordered_set<uint32_t> bucket = buckets_[bucket_idx];
   if(buckets_[bucket_idx].find(fingerprint) != buckets_[bucket_idx].end()) {
     buckets_[bucket_idx].erase(fingerprint);
     return true;
@@ -21,12 +21,12 @@ bool CuckooHashTable::Remove(uint16_t fingerprint, int bucket_idx) {
   }
 }
 
-bool CuckooHashTable::Contains(uint16_t fingerprint, int bucket_idx) {
+bool CuckooHashTable::Contains(uint32_t fingerprint, int bucket_idx) {
   return buckets_[bucket_idx].find(fingerprint) != buckets_[bucket_idx].end();
 }
 
-uint16_t CuckooHashTable::SwapEntries(uint16_t fingerprint, int bucket_idx) {
-  std::vector<uint16_t> bucket_cpy(buckets_[bucket_idx].size());
+uint32_t CuckooHashTable::SwapEntries(uint32_t fingerprint, int bucket_idx) {
+  std::vector<uint32_t> bucket_cpy(buckets_[bucket_idx].size());
   std::copy(buckets_[bucket_idx].begin(), buckets_[bucket_idx].end(), bucket_cpy.begin());
 
   // random generator
@@ -35,7 +35,7 @@ uint16_t CuckooHashTable::SwapEntries(uint16_t fingerprint, int bucket_idx) {
   std::uniform_int_distribution<int> uni(0, buckets_[bucket_idx].size() - 1); // guaranteed unbiased
   int swap_idx = uni(rng);
 
-  uint16_t old_entry = bucket_cpy[swap_idx];
+  uint32_t old_entry = bucket_cpy[swap_idx];
   buckets_[bucket_idx].erase(old_entry);
   buckets_[bucket_idx].insert(fingerprint);
   return old_entry;
